Random Numbers From ZINB Distribution

ZINB = zero inflated negative binomial.

rZINB(N, mu.nb, theta.nb, phi.zi)

Arguments

N
the number of random numbers to produce

mu.nb
the mean of the NB distribution

theta.nb
the dispersion parameter of the NB distribution, a single nonnegative numeric value or `NULL` (this refers to no overdispersion, thus a Poisson or ZIP distribution)

phi.zi
the probability of the non-zero valies in the ZI part

Value

A numeric vector of length `N` with random numbers

Examples

table(rZINB(100, 3, 2, 0.5))

#> 
#>  0  1  2  3  4  5  6  7  9 
#> 66  9  5  8  3  2  4  1  2 

table(rZINB(100, 3, NULL, 0.5))

#> 
#>  0  1  2  3  4  5  6  7 
#> 44  9 11 15 10  5  5  1